user projects query fn in profile page
